requirements.txt
dvc[s3]==3.54.0
mlflow==2.14.3
pandas==2.0.3
numpy==1.26.4
scikit-learn==1.5.1
xgboost==2.1.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
joblib==1.4.2
pyyaml==6.0.2
requests==2.32.3
loguru==0.7.2